Replaced Responses with overridable methods in AdviceTrait
