+++ src/main/java/org/zalando/problem/spring/web/advice/AdviceTrait.java
@@ -20,9 +20,13 @@ package org.zalando.problem.spring.web.advice;
  * #L%
  */
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.custom.CustomAdviceTrait;
 import org.zalando.problem.spring.web.advice.general.GeneralAdviceTrait;
@@ -31,18 +35,25 @@ import org.zalando.problem.spring.web.advice.io.IOAdviceTrait;
 import org.zalando.problem.spring.web.advice.routing.RoutingAdviceTrait;
 import org.zalando.problem.spring.web.advice.validation.ValidationAdviceTrait;
 
+import javax.ws.rs.core.Response;
+import java.util.Optional;
+import java.util.function.Function;
+
+import static java.util.function.Function.identity;
+import static org.zalando.problem.spring.web.advice.MediaTypes.determineContentType;
+
 /**
  * <p>
- *     Advice traits are simple interfaces that provide a single method with a default
- *     implementation. They are used to provide {@link ExceptionHandler} implementations to be used in
- *     {@link Controller Controllers} and/or in a {@link ControllerAdvice}. Clients can choose which traits they what to
- *     use à la carte.
+ * Advice traits are simple interfaces that provide a single method with a default
+ * implementation. They are used to provide {@link ExceptionHandler} implementations to be used in
+ * {@link Controller Controllers} and/or in a {@link ControllerAdvice}. Clients can choose which traits they what to
+ * use à la carte.
  * </p>
- *
+ * <p/>
  * <p>
- *     Advice traits are grouped in packages, based on they use cases. Every package has a composite advice trait that
- *     bundles all traits of that package. Additionally there is one {@link ProblemHandling major composite advice trait}
- *     that in turn bundles all other composites.
+ * Advice traits are grouped in packages, based on they use cases. Every package has a composite advice trait that
+ * bundles all traits of that package. Additionally there is one {@link ProblemHandling major composite advice trait}
+ * that in turn bundles all other composites.
  * </p>
  *
  * @see ControllerAdvice
@@ -59,5 +70,43 @@ import org.zalando.problem.spring.web.advice.validation.ValidationAdviceTrait;
  * @see ValidationAdviceTrait
  */
 public interface AdviceTrait {
+
+    default ResponseEntity<Problem> create(final Response.StatusType status, final Throwable throwable,
+            final NativeWebRequest request,
+            final Function<ResponseEntity.BodyBuilder, ResponseEntity.BodyBuilder> buildable) {
+        return create(status, throwable.getMessage(), request, buildable);
+    }
+
+    default ResponseEntity<Problem> create(final Response.StatusType status, final Throwable throwable,
+            final NativeWebRequest request) {
+        return create(status, throwable, request, identity());
+    }
+
+    default ResponseEntity<Problem> create(final Response.StatusType status, final String message,
+            final NativeWebRequest request,
+            final Function<ResponseEntity.BodyBuilder, ResponseEntity.BodyBuilder> buildable) {
+        return create(Problem.valueOf(status, message), request, buildable);
+    }
+
+    default ResponseEntity<Problem> create(final Response.StatusType status, final String message,
+            final NativeWebRequest request) {
+        return create(status, message, request, identity());
+    }
+
+    default ResponseEntity<Problem> create(final Problem problem, final NativeWebRequest request) {
+        return create(problem, request, identity());
+    }
+
+    default ResponseEntity<Problem> create(final Problem problem, final NativeWebRequest request,
+            final Function<ResponseEntity.BodyBuilder, ResponseEntity.BodyBuilder> buildable) {
+        final HttpStatus status = HttpStatus.valueOf(problem.getStatus().getStatusCode());
+        final ResponseEntity.BodyBuilder builder = buildable.apply(ResponseEntity.status(status));
+
+        final Optional<MediaType> contentType = determineContentType(request);
+        if (contentType.isPresent()) {
+            return builder.contentType(contentType.get()).body(problem);
+        }
+        return builder.body(null);
+    }
 
 }
