+++ src/main/java/org/zalando/problem/spring/web/advice/general/ThrowableAdviceTrait.java
@@ -25,11 +25,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.AdviceTrait;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.ws.rs.core.Response.Status;
-
-import static org.zalando.problem.spring.web.advice.Responses.create;
 
 /**
  * @see Throwable
@@ -42,7 +39,7 @@ public interface ThrowableAdviceTrait extends AdviceTrait {
     default ResponseEntity<Problem> handleThrowable(
             final Throwable throwable,
             final NativeWebRequest request) {
-        return Responses.create(Status.INTERNAL_SERVER_ERROR, throwable, request);
+        return create(Status.INTERNAL_SERVER_ERROR, throwable, request);
     }
 
 }
