+++ src/main/java/org/zalando/problem/spring/web/advice/io/MessageNotReadableAdviceTrait.java
@@ -26,7 +26,6 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.AdviceTrait;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.ws.rs.core.Response.Status;
 
@@ -42,7 +41,7 @@ public interface MessageNotReadableAdviceTrait extends AdviceTrait {
     default ResponseEntity<Problem> handleMessageNotReadableException(
             final HttpMessageNotReadableException exception,
             final NativeWebRequest request) {
-        return Responses.create(Status.BAD_REQUEST, exception, request);
+        return create(Status.BAD_REQUEST, exception, request);
     }
 
 }
