+++ src/main/java/org/zalando/problem/spring/web/advice/io/MultipartAdviceTrait.java
@@ -26,7 +26,6 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.multipart.MultipartException;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.AdviceTrait;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.ws.rs.core.Response.Status;
 
@@ -37,7 +36,7 @@ public interface MultipartAdviceTrait extends AdviceTrait {
     default ResponseEntity<Problem> handleMultipart(
             final MultipartException exception,
             final NativeWebRequest request) {
-        return Responses.create(Status.BAD_REQUEST, exception, request);
+        return create(Status.BAD_REQUEST, exception, request);
     }
 
 }
