+++ src/main/java/org/zalando/problem/spring/web/advice/io/TypeMistmatchAdviceTrait.java
@@ -26,7 +26,6 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.AdviceTrait;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.ws.rs.core.Response.Status;
 
@@ -40,6 +39,6 @@ public interface TypeMistmatchAdviceTrait extends AdviceTrait {
     default ResponseEntity<Problem> handleTypeMismatch(
             final TypeMismatchException exception,
             final NativeWebRequest request) {
-        return Responses.create(Status.BAD_REQUEST, exception, request);
+        return create(Status.BAD_REQUEST, exception, request);
     }
 }
