+++ src/main/java/org/zalando/problem/spring/web/advice/routing/NoSuchRequestHandlingMethodAdviceTrait.java
@@ -26,7 +26,6 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.servlet.mvc.multiaction.NoSuchRequestHandlingMethodException;
 import org.zalando.problem.Problem;
 import org.zalando.problem.spring.web.advice.AdviceTrait;
-import org.zalando.problem.spring.web.advice.Responses;
 
 import javax.ws.rs.core.Response.Status;
 
@@ -40,7 +39,7 @@ public interface NoSuchRequestHandlingMethodAdviceTrait extends AdviceTrait {
     default ResponseEntity<Problem> handleNoSuchRequestHandlingMethod(
             final NoSuchRequestHandlingMethodException exception,
             final NativeWebRequest request) {
-        return Responses.create(Status.NOT_FOUND, exception, request);
+        return create(Status.NOT_FOUND, exception, request);
     }
 
 }
