+++ src/test/java/org/zalando/problem/spring/web/advice/AdviceTraitTesting.java
@@ -1,0 +1,48 @@
+package org.zalando.problem.spring.web.advice;
+
+/*
+ * #%L
+ * problem-spring-web
+ * %%
+ * Copyright (C) 2015 Zalando SE
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.zalando.problem.ProblemModule;
+import org.zalando.problem.spring.web.advice.example.ExampleRestController;
+
+public interface AdviceTraitTesting {
+
+    default Object unit() {
+        return new ExceptionHandling();
+    }
+
+    default MockMvc mvc() {
+        final ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new Jdk8Module());
+        mapper.registerModule(new ProblemModule());
+
+        return MockMvcBuilders.standaloneSetup(new ExampleRestController())
+                .setControllerAdvice(unit())
+                .setMessageConverters(new MappingJackson2HttpMessageConverter(mapper))
+                .build();
+    }
+
+}
