+++ src/test/java/org/zalando/problem/spring/web/advice/general/UnsupportedOperationAdviceTraitTest.java
@@ -21,7 +21,7 @@ package org.zalando.problem.spring.web.advice.general;
  */
 
 import org.junit.Test;
-import org.zalando.problem.spring.web.advice.AdviceTraitTest;
+import org.zalando.problem.spring.web.advice.AdviceTraitTesting;
 
 import static org.hamcrest.Matchers.is;
 import static org.springframework.http.HttpMethod.GET;
@@ -30,7 +30,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public final class UnsupportedOperationAdviceTraitTest implements AdviceTraitTest {
+public final class UnsupportedOperationAdviceTraitTest implements AdviceTraitTesting {
 
     @Test
     public void unsupportedOperation() throws Exception {
