+++ src/test/java/org/zalando/problem/spring/web/advice/io/MessageNotReadableAdviceTraitTest.java
@@ -22,7 +22,7 @@ package org.zalando.problem.spring.web.advice.io;
 
 
 import org.junit.Test;
-import org.zalando.problem.spring.web.advice.AdviceTraitTest;
+import org.zalando.problem.spring.web.advice.AdviceTraitTesting;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
@@ -32,7 +32,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public final class MessageNotReadableAdviceTraitTest implements AdviceTraitTest {
+public final class MessageNotReadableAdviceTraitTest implements AdviceTraitTesting {
 
     @Test
     public void missingRequestBody() throws Exception {
