+++ src/test/java/org/zalando/problem/spring/web/advice/validation/ConstraintViolationAdviceTraitTest.java
@@ -21,7 +21,7 @@ package org.zalando.problem.spring.web.advice.validation;
  */
 
 import org.junit.Test;
-import org.zalando.problem.spring.web.advice.AdviceTraitTest;
+import org.zalando.problem.spring.web.advice.AdviceTraitTesting;
 
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
@@ -31,7 +31,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public final class ConstraintViolationAdviceTraitTest implements AdviceTraitTest {
+public final class ConstraintViolationAdviceTraitTest implements AdviceTraitTesting {
 
     @Test
     public void invalidRequestParam() throws Exception {
